chore: move Option to dynamic
